Move `WaitingSender` to `inbox` module
