+++ src/inbox.rs
@@ -6,8 +6,11 @@ pub mod tx;
 
 use std::cmp::Ordering;
 use std::collections::{BinaryHeap, VecDeque};
+use std::future::Future;
+use std::pin::Pin;
 use std::sync::atomic::AtomicUsize;
 use std::sync::{atomic, Arc, Mutex, Weak};
+use std::task::{Context, Poll, Waker};
 use std::{cmp, mem};
 
 use event_listener::{Event, EventListener};
@@ -16,8 +19,8 @@ pub use tx::{SendFuture, Sender};
 
 use crate::envelope::{BroadcastEnvelope, MessageEnvelope, Shutdown};
 use crate::inbox::rx::{RxStrong, WaitingReceiver};
-use crate::inbox::tx::{TxStrong, WaitingSender};
-use crate::Actor;
+use crate::inbox::tx::TxStrong;
+use crate::{Actor, Error};
 
 type Spinlock<T> = spin::Mutex<T>;
 pub type MessageToOneActor<A> = Box<dyn MessageEnvelope<Actor = A>>;
@@ -426,6 +429,76 @@ impl<A> ChanInner<A> {
             if let Some(tx) = self.waiting_senders.remove(pos).unwrap().upgrade() {
                 return Some(tx.lock().fulfill());
             }
+        }
+    }
+}
+
+pub enum WaitingSender<A> {
+    Active {
+        waker: Option<Waker>,
+        message: SentMessage<A>,
+    },
+    Delivered,
+    Closed,
+}
+
+impl<A> WaitingSender<A> {
+    pub fn new(message: SentMessage<A>) -> Arc<Spinlock<Self>> {
+        let sender = WaitingSender::Active {
+            waker: None,
+            message,
+        };
+        Arc::new(Spinlock::new(sender))
+    }
+
+    pub fn peek(&self) -> &SentMessage<A> {
+        match self {
+            WaitingSender::Active { message, .. } => message,
+            _ => panic!("WaitingSender should have message"),
+        }
+    }
+
+    pub fn fulfill(&mut self) -> SentMessage<A> {
+        match mem::replace(self, Self::Delivered) {
+            WaitingSender::Active { mut waker, message } => {
+                if let Some(waker) = waker.take() {
+                    waker.wake();
+                }
+
+                message
+            }
+            WaitingSender::Delivered | WaitingSender::Closed => {
+                panic!("WaitingSender is already fulfilled or closed")
+            }
+        }
+    }
+
+    /// Mark this [`WaitingSender`] as closed.
+    ///
+    /// Should be called when the last [`Receiver`](crate::inbox::Receiver) goes away.
+    pub fn set_closed(&mut self) {
+        if let WaitingSender::Active {
+            waker: Some(waker), ..
+        } = mem::replace(self, Self::Closed)
+        {
+            waker.wake();
+        }
+    }
+}
+
+impl<A> Future for WaitingSender<A> {
+    type Output = Result<(), Error>;
+
+    fn poll(self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Self::Output> {
+        let this = self.get_mut();
+
+        match this {
+            WaitingSender::Active { waker, .. } => {
+                *waker = Some(cx.waker().clone());
+                Poll::Pending
+            }
+            WaitingSender::Delivered => Poll::Ready(Ok(())),
+            WaitingSender::Closed => Poll::Ready(Err(Error::Disconnected)),
         }
     }
 }
