+++ src/inbox/tx.rs
@@ -3,7 +3,7 @@ use std::future::Future;
 use std::mem;
 use std::pin::Pin;
 use std::sync::{atomic, Arc};
-use std::task::{Context, Poll, Waker};
+use std::task::{Context, Poll};
 
 use event_listener::EventListener;
 use futures_core::FusedFuture;
@@ -179,76 +179,6 @@ impl<A, Rc: TxRefCounter> Future for SendFuture<A, Rc> {
                 }
                 SendFuture::Complete => return Poll::Pending,
             }
-        }
-    }
-}
-
-pub enum WaitingSender<A> {
-    Active {
-        waker: Option<Waker>,
-        message: SentMessage<A>,
-    },
-    Delivered,
-    Closed,
-}
-
-impl<A> WaitingSender<A> {
-    pub fn new(message: SentMessage<A>) -> Arc<Spinlock<Self>> {
-        let sender = WaitingSender::Active {
-            waker: None,
-            message,
-        };
-        Arc::new(Spinlock::new(sender))
-    }
-
-    pub fn peek(&self) -> &SentMessage<A> {
-        match self {
-            WaitingSender::Active { message, .. } => message,
-            _ => panic!("WaitingSender should have message"),
-        }
-    }
-
-    pub fn fulfill(&mut self) -> SentMessage<A> {
-        match mem::replace(self, Self::Delivered) {
-            WaitingSender::Active { mut waker, message } => {
-                if let Some(waker) = waker.take() {
-                    waker.wake();
-                }
-
-                message
-            }
-            WaitingSender::Delivered | WaitingSender::Closed => {
-                panic!("WaitingSender is already fulfilled or closed")
-            }
-        }
-    }
-
-    /// Mark this [`WaitingSender`] as closed.
-    ///
-    /// Should be called when the last [`Receiver`](crate::inbox::Receiver) goes away.
-    pub fn set_closed(&mut self) {
-        if let WaitingSender::Active {
-            waker: Some(waker), ..
-        } = mem::replace(self, Self::Closed)
-        {
-            waker.wake();
-        }
-    }
-}
-
-impl<A> Future for WaitingSender<A> {
-    type Output = Result<(), Error>;
-
-    fn poll(self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Self::Output> {
-        let this = self.get_mut();
-
-        match this {
-            WaitingSender::Active { waker, .. } => {
-                *waker = Some(cx.waker().clone());
-                Poll::Pending
-            }
-            WaitingSender::Delivered => Poll::Ready(Ok(())),
-            WaitingSender::Closed => Poll::Ready(Err(Error::Disconnected)),
         }
     }
 }
